Implemented clicking
When the program first runs, it reads from the score file to
keep track of the players score during their play time.
Added a two buttons and a label. The first button, when clicked,
adds 1 to the score and prints the new score to the label.
The second button re-writes the score file to show the new score
then closes the file and the program.